MAXZOOM to 16, Hilbert utility use long instead of int

+++ planetiler-core/src/main/java/com/onthegomap/planetiler/collection/FeatureGroup.java
@@ -167,8 +167,8 @@ public final class FeatureGroup implements Iterable<FeatureGroup.TileFeatures>, 
    * Encode key by {@code tile} asc, {@code layer} asc, {@code sortKey} asc with an extra bit to indicate whether the
    * value contains grouping information.
    */
-  static long encodeKey(int tile, byte layer, int sortKey, boolean hasGroup) {
-    return ((long) tile << 32L) | ((long) (layer & 0xff) << 24L) | (((sortKey - SORT_KEY_MIN) & SORT_KEY_MASK) << 1L) |
+  static long encodeKey(long tile, byte layer, int sortKey, boolean hasGroup) {
+    return (tile << 32L) | ((long) (layer & 0xff) << 24L) | (((sortKey - SORT_KEY_MIN) & SORT_KEY_MASK) << 1L) |
       (hasGroup ? 1 : 0);
   }
 
@@ -176,8 +176,8 @@ public final class FeatureGroup implements Iterable<FeatureGroup.TileFeatures>, 
     return (key & 1) == 1;
   }
 
-  static int extractTileFromKey(long key) {
-    return (int) (key >> 32L);
+  static long extractTileFromKey(long key) {
+    return key >> 32L;
   }
 
   static byte extractLayerIdFromKey(long key) {
@@ -342,7 +342,7 @@ public final class FeatureGroup implements Iterable<FeatureGroup.TileFeatures>, 
     SortableFeature firstFeature = entries.next();
     return new Iterator<>() {
       private SortableFeature lastFeature = firstFeature;
-      private int lastTileId = extractTileFromKey(firstFeature.key());
+      private long lastTileId = extractTileFromKey(firstFeature.key());
 
       @Override
       public boolean hasNext() {
@@ -353,7 +353,7 @@ public final class FeatureGroup implements Iterable<FeatureGroup.TileFeatures>, 
       public TileFeatures next() {
         TileFeatures result = new TileFeatures(lastTileId);
         result.add(lastFeature);
-        int lastTile = lastTileId;
+        long lastTile = lastTileId;
 
         while (entries.hasNext()) {
           SortableFeature next = entries.next();
@@ -408,7 +408,7 @@ public final class FeatureGroup implements Iterable<FeatureGroup.TileFeatures>, 
     private LongLongHashMap counts = null;
     private byte lastLayer = Byte.MAX_VALUE;
 
-    private TileFeatures(int lastTileId) {
+    private TileFeatures(long lastTileId) {
       this.tileCoord = tileOrder.decode(lastTileId);
     }
